Specify target repo name (#9)
* add target repo name

This addition of the user-specified target repo name is intended to resolve issues with running the tool when the source repo has already been cloned into the organization or there is a separate naming collision. This logic is also a prerequisite for multi-phase audits which potentially require separate setup for multiple different commit hashes of the same repository. It is easiest and cleanest to simply create a separate internal repo that is ossified after each engagement, rather than attempting to update and re-use it for each new phase.

* add additional context to prompt